feat: use decimal field for bmi, not positive small integer field

+++ dialog_backend/data_tracking/models.py
@@ -102,7 +102,9 @@ class WeeklyLog(AutoDateMixin):
         verbose_name='Вес',
     )
 
-    bmi = models.PositiveSmallIntegerField(
+    bmi = models.DecimalField(
+        max_digits=5,
+        decimal_places=1,
         verbose_name='Индекс массы тела',
     )
